+++ src/main/webapp/assets/css/style.css
@@ -35,7 +35,7 @@
     text-align: center;
 }
 img{
-	max-width: 100px;
+	max-width: 100%;
     height: 350px;
