+++ src/main/webapp/header.jsp
@@ -1,6 +1,8 @@
 <link rel="stylesheet" href="assets/css/bootstrap.min.css">
 <link rel="stylesheet" href="assets/css/fontawesome.min.css">
 <link rel="stylesheet" href="assets/css/style.css">
+<%   String  LoggedUser= (String)session.getAttribute("Logged_in_User");
+%>
 
 <header>
 <nav class="navbar navbar-expand-sm navbar-dark bg-dark">
@@ -26,12 +28,23 @@
       </li>
     </ul>
      <ul class="navbar-nav ml-auto mt-2 mt-lg-0">
+     <% if (LoggedUser==null) { %>
       <li class="nav-item active">
         <a class="nav-link" href="LoginPage.jsp">Login</a>
       </li>
       <li class="nav-item">
         <a class="nav-link" href="register.jsp">Register</a>
       </li>
+      <% }  else {%>
+      <li class="nav-item">
+        <a class="nav-link" href="#">Welcome  <%=LoggedUser%></a>
+       
+      </li>
+       <li class="nav-item">
+        <a class="nav-link" href="LogoutServlet">Logout</a>
+       
+      </li>
+      <%} %>
       </ul>
    
   </div>
